requirements.txt
confluent-kafka
configparser
flask
python-dotenv
cryptography